[frontend] refactor add react route
